Began Analysis
I started doing the analysis with knn classification, adding predictors one at a time to see what combination of predictors leads to highest accuracy - as we had mentioned we would do in our proposal. The accuracy is very low - this may be due to how unbalanced the distribution of quality is in our dataset.